Restructure and move drawing functionality
Move logic for drawing out of component/ directory into drawing/. Refactor
drawcontrol reducer and move livestroke update into drawing/.
