+++ src/drawing/simplify.js
